Rename flag component files
