+++ src/flags/cn.jsx
