+++ src/flags/fr.jsx
