+++ src/flags/no.jsx
